Semester - 3/Course 06 Database Management (3 Credits)/Week 04 Connection with Python & Project/Module 15 MySQL Using Python/create_db.py
import mysql.connector

mydb = mysql.connector.connect(
    host="localhost",
    user="root",
    password="*$#Course@1998"
)

# Create a database
mycursor = mydb.cursor()

db_name = "mydatabase"
sql_command = "CREATE DATABASE " + db_name

mycursor.execute(sql_command)

