Semester - 3/Course 06 Database Management (3 Credits)/Week 04 Connection with Python & Project/Module 15 MySQL Using Python/requirements.txt
mysql-connector-python==8.1.0
protobuf==4.21.12
